Update DOMException name: InUseAttributeError https://bugs.webkit.org/show_bug.cgi?id=102141
Reviewed by Ojan Vafai.

Patch 10 of 25 to update DOMException name to match the spec and Firefox.

The name for this is not in the spec but the case was selected to match
Firefox. http://mxr.mozilla.org/mozilla-central/source/dom/base/domerr.msg#18

INUSE_ATTRIBUTE_ERR is historical and not used in any spec or any of our tests.

* dom/DOMCoreException.cpp:
* dom/ExceptionCode.h:


Identifier: 120383@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@134659 268f45cc-cd09-0410-ab3c-d52691b4dbfc