WebSocket's MessageEvent.origin attribute is an empty string https://bugs.webkit.org/show_bug.cgi?id=102889
Patch by Christophe Dumez <christophe.dumez@intel.com> on 2012-11-23
Reviewed by Kentaro Hara.

Source/WebCore:

Fix WebSocket implementation so that the message event's origin
attribute is properly set to the Unicode serialization of the
origin of the URL that was passed to the WebSocket object's
constructor, instead of an empty string.

The new behavior is according to the specification:
http://www.whatwg.org/specs/web-apps/current-work/multipage/network.html#feedback-from-the-protocol

It also matches the behavior of Firefox.

Tests: http/tests/websocket/tests/hybi/send-blob-onmessage-origin.html
       http/tests/websocket/tests/hybi/send-onmessage-origin.html

* Modules/websockets/WebSocket.cpp:
(WebCore::WebSocket::didReceiveMessage):
(WebCore::WebSocket::didReceiveBinaryData):
* dom/MessageEvent.cpp:
(WebCore::MessageEvent::MessageEvent):
* dom/MessageEvent.h:
(WebCore::MessageEvent::create):
(MessageEvent):

LayoutTests:

Add layout tests to check that the message event's origin attribute
is set to the Unicode serialization of the origin of the URL that
was passed to the WebSocket object's constructor, as per the
WebSocket specification.

* http/tests/websocket/tests/hybi/send-blob-onmessage-origin-expected.txt: Added.
* http/tests/websocket/tests/hybi/send-blob-onmessage-origin.html: Added.
* http/tests/websocket/tests/hybi/send-onmessage-origin-expected.txt: Added.
* http/tests/websocket/tests/hybi/send-onmessage-origin.html: Added.

Identifier: 121248@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@135587 268f45cc-cd09-0410-ab3c-d52691b4dbfc