Page background color bleeds through inner edge of div border with rounded edges https://bugs.webkit.org/show_bug.cgi?id=103089
Reviewed by Stephen White.

Source/WebCore:

Test that verifies whether a RenderBoxEdge is more than 2 pixels wide
was applying the content scale factor backwards, so the
obscuresBackgroundEdge was falsely passing for downsized boxes.

Test: fast/backgrounds/gradient-background-leakage-2.html

* rendering/RenderBoxModelObject.cpp:
(WebCore::BorderEdge::obscuresBackgroundEdge):

LayoutTests:

New layout test to verify that there is no color bleeding between
a div element's border and background when there are rounded corners
an a scale that reduces a thick border down to a width of less than
two pixels.

* fast/backgrounds/gradient-background-leakage-2-expected.txt: Added.
* fast/backgrounds/gradient-background-leakage-2.html: Added.
* platform/chromium/TestExpectations:


Identifier: 121286@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@135625 268f45cc-cd09-0410-ab3c-d52691b4dbfc