DataLog to a file should work if there are multiple processes using WTF https://bugs.webkit.org/show_bug.cgi?id=103323
Reviewed by Mark Hahnenberg.

Whereas before DataLog would open a file with the name you specified, now it'll open a file with the
name plus the PID appended to it. So if you are dealing with multiple processes running with DataLog
to a file enabled, you'll get multiple separate log files.

* wtf/DataLog.cpp:
(WTF::initializeLogFileOnce):



Identifier: 121440@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@135790 268f45cc-cd09-0410-ab3c-d52691b4dbfc