Layout Test svg/animations/mpath-remove-from-dependents-on-delete-crash.html is flaky https://bugs.webkit.org/show_bug.cgi?id=103375
Unreviewed test update.

According to the original bug, at least 4 iterations were required to
crash (without the fix). To make the test faster, I've reduced the max
iteration count to 5, rather than 10. This may not fix the timeoout
issue because it might be a real timeout, not just a slow test.

* svg/animations/mpath-remove-from-dependents-on-delete-crash.html:


Identifier: 121655@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136024 268f45cc-cd09-0410-ab3c-d52691b4dbfc